LayoutTests/gamepad/gamepad-vibration-document-no-longer-fully-active-expected.txt

Tests that the [playingEffectPromise] gets resolved with 'preempted' when the document is no longer fully active

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Starting dual-rumble effect
* Detach iframe
PASS result is "preempted"
PASS successfullyParsed is true

TEST COMPLETE
